Use generated Event Model

+++ android/src/test/java/com/thebluealliance/androidclient/datafeed/maps/WeekEventsExtractorTest.java
@@ -38,7 +38,7 @@ public class WeekEventsExtractorTest {
         assertEquals(events.size(), 21);
         for (int i = 0; i < events.size(); i++) {
             Event event = events.get(i);
-            assertEquals(event.getCompetitionWeek(), mWeek);
+            assertEquals(event.getCompetitionWeek(), (Integer)mWeek);
         }
     }
 }
